+++ kernel/src/elos/kernel/exec/syscall_handler.c
@@ -743,7 +743,7 @@ u64 EXEC_syscall_handler(u64 arg0, u64 arg1, u64 arg2, u64 arg3, u64 arg4, u64 a
             
             write_cr3((u64)g_kernelPageTable);
 
-            u64 wholeBufferSize;
+            u32 wholeBufferSize;
             ELOS_UserEventBuffer* tmp_buffer;
             bool result = EVE_request_user_event_buffer(minimumEvents, &tmp_buffer, &wholeBufferSize);
